replace custom handling with express-async-errors and moved error handler to utilities.
