+++ server/package.json
@@ -27,6 +27,7 @@
     "cors": "^2.8.5",
     "dotenv-safe": "^8.2.0",
     "express": "^4.18.2",
+    "express-async-errors": "^3.1.1",
     "express-validator": "^7.0.1",
     "nodemon": "^3.0.3",
     "pg": "^8.11.3",
