+++ server/src/api/root_api.ts
@@ -1,6 +1,5 @@
 import {Request, Response, Router} from 'express';
 import {ResponseHandler} from './response_handler';
-import {ErrorHandler} from './error_handler';
 
 /**
  * This class represents the default api route reached by not specifying a resource or action
@@ -9,12 +8,10 @@ class RootApi {
   /**
    * Dependencies needed for api operation are injected via this constructor.
    * @param responseHandler a handler used for formatting and sending HTTP responses.
-   * @param errorHandler a handler used for enabling error handling.
    * @param router the express route associated with this class.
    */
   constructor(
     private responseHandler: ResponseHandler,
-    private errorHandler: ErrorHandler,
     private router: Router
   ) {}
 
@@ -25,7 +22,6 @@ class RootApi {
   async setupRequestHandling() {
     this.router.get(
       '/',
-      this.errorHandler.asyncErrorWrapper(
         async (req: Request, res: Response) => {
           const data = {message: 'API is up and running!'};
           const httpStatusCode = 200;
@@ -36,7 +32,6 @@ class RootApi {
             false
           );
         }
-      )
     );
   }
 }
